+++ tests/utils.py
@@ -32,6 +32,7 @@ from pyramid.config import Configurator
 from pyramid.httpexceptions import HTTPException, HTTPNotFound, HTTPUnprocessableEntity
 from pyramid.registry import Registry
 from pyramid.testing import DummyRequest
+from pytest_server_fixtures.http import HTTPTestServer, SimpleHTTPTestServer
 from requests import Response
 from webtest import TestApp, TestResponse
 
@@ -843,6 +844,30 @@ def mocked_dir_listing(local_directory,             # type: str
     </html>
     """)
     return dir_html
+
+
+class FileServer(SimpleHTTPTestServer):
+    """
+    Generate a file server that can host files under :attr:`document_root`.
+
+    Contrary to :func:`mocked_file_server` where requests are captured and redirected to the corresponding files,
+    this server receives *real* requests (via a socket) and returns matched files. This is particularly important
+    during tests that call subprocesses independently of the main Python test process
+    (e.g.: :term:`CWL` ``CommandLineTool` for a ``Builtin`` :term:`Process`), because mocks applied in the main
+    process are not reflected in the other ones.
+
+    .. warning::
+        This server takes more time to start than usual mocks. Use it sparingly, and consider maintaining a single
+         instance over multiple tests of a complete test suite rather than recreating a server for each test.
+    """
+    def __init__(self):
+        self._port = self.get_port()
+        self._uri = f"http://0.0.0.0:{self._port}"
+
+        # purposely call 'HTTPTestServer' instead of 'SimpleHTTPTestServer' to enforce the URI
+        # otherwise, 'socket.gethostname()' is used (machine name), and the obtained URI fail our schema validation
+        HTTPTestServer.__init__(self, hostname="0.0.0.0", port=self._port, uri=self._uri)
+        self.cwd = self.document_root
 
 
 def mocked_file_server(directory,                   # type: str