+++ weaver/processes/builtin/file2string_array.py
@@ -9,16 +9,26 @@ import os
 import sys
 
 CUR_DIR = os.path.abspath(os.path.dirname(__file__))
+sys.path.insert(0, CUR_DIR)
+# root to allow 'from weaver import <...>'
+sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.dirname(CUR_DIR))))
+
+# place weaver specific imports after sys path fixing to ensure they are found from external call
+# pylint: disable=C0413,wrong-import-order
+from weaver import WEAVER_ROOT_DIR  # isort:skip # noqa: E402
+from weaver.processes.builtin.utils import validate_file_reference  # isort:skip # noqa: E402
 
 PACKAGE_NAME = os.path.split(os.path.splitext(__file__)[0])[-1]
+PACKAGE_BASE = __file__.split(WEAVER_ROOT_DIR.rstrip("/") + "/")[-1].rsplit(PACKAGE_NAME)[0]
+PACKAGE_MODULE = f"{PACKAGE_BASE}{PACKAGE_NAME}".replace("/", ".")
 
 # setup logger since it is not run from the main 'weaver' app
-LOGGER = logging.getLogger(__name__)
+LOGGER = logging.getLogger(PACKAGE_MODULE)
 LOGGER.addHandler(logging.StreamHandler(sys.stdout))
 LOGGER.setLevel(logging.INFO)
 
 # process details
-__version__ = "1.2"
+__version__ = "1.3"
 __title__ = "File to String-Array"
 __abstract__ = __doc__  # NOTE: '__doc__' is fetched directly, this is mostly to be informative
 
@@ -26,8 +36,9 @@ OUTPUT_CWL_JSON = "cwl.output.json"
 
 
 def process(input_file, output_dir):
-    # type: (argparse.FileType, str) -> None
+    # type: (str, str) -> None
     LOGGER.info("Got arguments: input_file=%s output_dir=%s", input_file, output_dir)
+    validate_file_reference(input_file)
     output_data = {"output": [input_file]}
     with open(os.path.join(output_dir, OUTPUT_CWL_JSON), mode="w", encoding="utf-8") as file:
         return json.dump(output_data, file)