+++ weaver/processes/builtin/jsonarray2netcdf.py
@@ -17,20 +17,24 @@ sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.dirname(CUR_DIR))))
 
 # place weaver specific imports after sys path fixing to ensure they are found from external call
 # pylint: disable=C0413,wrong-import-order
-from weaver.processes.builtin.utils import is_netcdf_url  # isort:skip # noqa: E402
-from weaver.utils import fetch_file  # isort:skip # noqa: E402
+from weaver import WEAVER_ROOT_DIR  # isort:skip # noqa: E402
+from weaver.formats import repr_json  # isort:skip # noqa: E402
+from weaver.processes.builtin.utils import is_netcdf_url, validate_file_reference  # isort:skip # noqa: E402
+from weaver.utils import fetch_file, get_secure_path  # isort:skip # noqa: E402
 
 PACKAGE_NAME = os.path.split(os.path.splitext(__file__)[0])[-1]
+PACKAGE_BASE = __file__.split(WEAVER_ROOT_DIR.rstrip("/") + "/")[-1].rsplit(PACKAGE_NAME)[0]
+PACKAGE_MODULE = f"{PACKAGE_BASE}{PACKAGE_NAME}".replace("/", ".")
 
 # setup logger since it is not run from the main 'weaver' app
-LOGGER = logging.getLogger(PACKAGE_NAME)
+LOGGER = logging.getLogger(PACKAGE_MODULE)
 _handler = logging.StreamHandler(sys.stdout)  # noqa
 _handler.setFormatter(logging.Formatter(fmt="[%(name)s] %(levelname)-8s %(message)s"))
 LOGGER.addHandler(_handler)
-LOGGER.setLevel(logging.DEBUG)
+LOGGER.setLevel(logging.INFO)
 
 # process details
-__version__ = "2.2"
+__version__ = "2.3"
 __title__ = "JSON array to NetCDF"
 __abstract__ = __doc__  # NOTE: '__doc__' is fetched directly, this is mostly to be informative
 
@@ -38,24 +42,31 @@ __abstract__ = __doc__  # NOTE: '__doc__' is fetched directly, this is mostly to
 def j2n(json_reference, output_dir):
     # type: (str, str) -> None
     LOGGER.info("Process '%s' execution starting...", PACKAGE_NAME)
-    LOGGER.debug("Process '%s' output directory: [%s].", PACKAGE_NAME, output_dir)
+    LOGGER.info("Process '%s' output directory: [%s].", PACKAGE_NAME, output_dir)
     try:
         if not os.path.isdir(output_dir):
             raise ValueError(f"Output dir [{output_dir}] does not exist.")
         with TemporaryDirectory(prefix=f"wps_process_{PACKAGE_NAME}_") as tmp_dir:
-            LOGGER.debug("Fetching JSON file: [%s]", json_reference)
+            LOGGER.info("Verify URL reference: [%s]", json_reference)
+            validate_file_reference(json_reference)
+            LOGGER.info("Fetching JSON file: [%s]", json_reference)
             json_path = fetch_file(json_reference, tmp_dir, timeout=10, retry=3)
-            LOGGER.debug("Reading JSON file: [%s]", json_path)
+            json_path = get_secure_path(json_path)
+            LOGGER.info("Reading JSON file: [%s]", json_path)
             with open(json_path, mode="r", encoding="utf-8") as json_file:
                 json_content = json.load(json_file)
-            if not isinstance(json_content, list) or any(not is_netcdf_url(f) for f in json_content):
+            if not isinstance(json_content, list) or any(not isinstance(item, str) for item in json_content):
                 LOGGER.error("Invalid JSON: [%s]", json_content)
                 raise ValueError("Invalid JSON file format, expected a plain array of NetCDF file URL strings.")
-            LOGGER.debug("Parsing JSON file references.")
+            LOGGER.info("Parsing JSON file references from file contents:\n%s", repr_json(json_content))
             for file_url in json_content:
-                LOGGER.debug("Fetching NetCDF reference from JSON file: [%s]", file_url)
+                LOGGER.info("Validate NetCDF reference from JSON file: [%s]", file_url)
+                validate_file_reference(file_url)
+                if not is_netcdf_url(file_url):
+                    raise ValueError(f"Invalid file format for [{file_url}], expected a NetCDF file URL.")
+                LOGGER.info("Fetching NetCDF reference from JSON file: [%s]", file_url)
                 fetched_nc = fetch_file(file_url, output_dir, timeout=10, retry=3)
-                LOGGER.debug("Fetched NetCDF output location: [%s]", fetched_nc)
+                LOGGER.info("Fetched NetCDF output location: [%s]", fetched_nc)
     except Exception as exc:
         LOGGER.error("Process '%s' raised an exception: [%s]", PACKAGE_NAME, exc)
         raise